+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `Array#key_map` and `Array#dig_map` for mapping over `Hash`
+- Added `with_index:` keyword argument to `Array#join_map`. Defaults to `false`
 
 ### Changed